Move mock into its own file.
