+++ tests/lib/ocpp/v201/test_charge_point.cpp
@@ -5,6 +5,7 @@
 #include "everest/logging.hpp"
 #include "evse_security_mock.hpp"
 #include "lib/ocpp/common/database_testing_utils.hpp"
+#include "message_queue_mock.hpp"
 #include "ocpp/common/call_types.hpp"
 #include "ocpp/common/message_queue.hpp"
 #include "ocpp/v201/charge_point.hpp"
@@ -130,39 +131,6 @@ public:
         auto database_connection = std::make_unique<common::DatabaseConnection>(fs::path("/tmp/ocpp201") / "cp.db");
         return std::make_shared<DatabaseHandler>(std::move(database_connection), MIGRATION_FILES_LOCATION_V201);
     }
-
-    class MessageQueueMock : public MessageQueueInterface<v201::MessageType, MessageQueue<v201::MessageType>> {
-    public:
-        MOCK_METHOD(void, start, ());
-        MOCK_METHOD(void, reset_next_message_to_send, ());
-        MOCK_METHOD(void, get_persisted_messages_from_db, (bool ignore_security_event_notifications));
-        MOCK_METHOD(void, push, (const json& message, const bool stall_until_accepted));
-        MOCK_METHOD(void, push, (CallError call_error));
-        MOCK_METHOD(EnhancedMessage<v201::MessageType>, receive, (std::string_view message));
-        MOCK_METHOD(void, reset_in_flight, ());
-        MOCK_METHOD(void, handle_call_result, (EnhancedMessage<v201::MessageType> & enhanced_message));
-        MOCK_METHOD(void, handle_timeout_or_callerror,
-                    (const std::optional<EnhancedMessage<v201::MessageType>>& enhanced_message_opt));
-        MOCK_METHOD(void, stop, ());
-        MOCK_METHOD(void, pause, ());
-        MOCK_METHOD(void, resume, (std::chrono::seconds delay_on_reconnect));
-        MOCK_METHOD(void, set_registration_status_accepted, ());
-        MOCK_METHOD(bool, is_transaction_message_queue_empty, ());
-        MOCK_METHOD(bool, contains_transaction_messages, (const CiString<36> transaction_id));
-        MOCK_METHOD(bool, contains_stop_transaction_message, (const int32_t transaction_id));
-        MOCK_METHOD(void, update_transaction_message_attempts, (const int transaction_message_attempts));
-        MOCK_METHOD(void, update_transaction_message_retry_interval, (const int transaction_message_retry_interval));
-        MOCK_METHOD(void, update_message_timeout, (const int timeout));
-        MOCK_METHOD(MessageId, createMessageId, ());
-        MOCK_METHOD(void, add_stopped_transaction_id,
-                    (std::string stop_transaction_message_id, int32_t transaction_id));
-        MOCK_METHOD(void, add_meter_value_message_id,
-                    (const std::string& start_transaction_message_id, const std::string& meter_value_message_id));
-        MOCK_METHOD(void, notify_start_transaction_handled,
-                    (const std::string& start_transaction_message_id, const int32_t transaction_id));
-        MOCK_METHOD(v201::MessageType, string_to_messagetype, (const std::string& s));
-        MOCK_METHOD(std::string, messagetype_to_string, (v201::MessageType m));
-    };
 
     std::shared_ptr<MessageQueueMock> create_message_queue(std::shared_ptr<DatabaseHandler>& database_handler) {
         return std::make_shared<MessageQueueMock>();
